+++ websites/admin/common.inc.php
@@ -31,13 +31,6 @@ function get_subtype_name($subtype){
 	}
 }
 
-function get_hentai_slug($series){
-	if ($series['rating']=='XXX') {
-		return "hentai/";
-	}
-	return "";
-}
-
 function get_fansub_preposition_name($text){
 	$first = substr($text, 0, 1);
 	if (($first == 'A' || $first == 'E' || $first == 'I' || $first == 'O' || $first == 'U') && substr($text, 0, 4)!='One '){ //Ugly...