+++ websites/admin/version_links.php
@@ -70,6 +70,9 @@ if (!empty($_SESSION['username']) && !empty($_SESSION['admin_level']) && $_SESSI
 		array_push($episodes, $rowe);
 	}
 	mysqli_free_result($resulte);
+	if ($series['rating']=='XXX') {
+		$link_url = str_replace(MAIN_DOMAIN, HENTAI_DOMAIN, $link_url);
+	}
 ?>
 		<div class="container d-flex justify-content-center p-4">
 			<div class="card w-100">
@@ -79,7 +82,7 @@ if (!empty($_SESSION['username']) && !empty($_SESSION['admin_level']) && $_SESSI
 					<p class="text-center">Aquests són els enllaços de la versió de "<b><?php echo htmlspecialchars($row['series_name']); ?></b>" feta per <?php echo htmlspecialchars($row['fansub_name']); ?>.</p>
 					<hr>
 					<div class="text-center">
-						<button onclick="copyToClipboard('<?php echo $link_url.'/'.get_hentai_slug($series).$series['slug']; ?>?v=<?php echo $_GET['id']; ?>', $(this));" class="btn btn-primary"><span class="fa fa-clipboard pe-2"></span>Copia l’enllaç a la fitxa de la versió</button>
+						<button onclick="copyToClipboard('<?php echo $link_url.'/'.$series['slug']; ?>?v=<?php echo $_GET['id']; ?>', $(this));" class="btn btn-primary"><span class="fa fa-clipboard pe-2"></span>Copia l’enllaç a la fitxa de la versió</button>
 					</div>
 				</article>
 			</div>
@@ -106,7 +109,7 @@ if (!empty($_SESSION['username']) && !empty($_SESSION['admin_level']) && $_SESSI
 ?>
 						<tr>
 							<td style="width: 70%;"><strong><?php echo !empty($division['name']) ? $division['name'] : $divisions_name." ".$division['number']; ?></strong></td>
-							<td class="text-center"><button onclick="copyToClipboard('<?php echo $link_url.'/'.get_hentai_slug($series).$series['slug'].'?v='.$_GET['id'].'#'.$divisions_anchor.'-'.$division['number']; ?>', $(this));" class="btn btn-sm btn-primary"><span class="fa fa-clipboard pe-2"></span>Copia l’enllaç</button></td>
+							<td class="text-center"><button onclick="copyToClipboard('<?php echo $link_url.'/'.$series['slug'].'?v='.$_GET['id'].'#'.$divisions_anchor.'-'.$division['number']; ?>', $(this));" class="btn btn-sm btn-primary"><span class="fa fa-clipboard pe-2"></span>Copia l’enllaç</button></td>
 						</tr>
 <?php
 		}